# mr/models/versioned_model.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from mr.models.base_model import Model
from mr.models.environment import Environment
from mr import util

from mr.models.environment import Environment  # noqa: E501

class VersionedModel(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, id=None, version=None, tags=None, labels=None, environments=None, created_date=None, updated_date=None):  # noqa: E501
        """VersionedModel - a model defined in OpenAPI

        :param name: The name of this VersionedModel.  # noqa: E501
        :type name: str
        :param id: The id of this VersionedModel.  # noqa: E501
        :type id: str
        :param version: The version of this VersionedModel.  # noqa: E501
        :type version: str
        :param tags: The tags of this VersionedModel.  # noqa: E501
        :type tags: List[str]
        :param labels: The labels of this VersionedModel.  # noqa: E501
        :type labels: List[str]
        :param environments: The environments of this VersionedModel.  # noqa: E501
        :type environments: List[Environment]
        :param created_date: The created_date of this VersionedModel.  # noqa: E501
        :type created_date: datetime
        :param updated_date: The updated_date of this VersionedModel.  # noqa: E501
        :type updated_date: datetime
        """
        self.openapi_types = {
            'name': str,
            'id': str,
            'version': str,
            'tags': List[str],
            'labels': List[str],
            'environments': List[Environment],
            'created_date': datetime,
            'updated_date': datetime
        }

        self.attribute_map = {
            'name': 'name',
            'id': 'id',
            'version': 'version',
            'tags': 'tags',
            'labels': 'labels',
            'environments': 'environments',
            'created_date': 'created date',
            'updated_date': 'updated date'
        }

        self._name = name
        self._id = id
        self._version = version
        self._tags = tags
        self._labels = labels
        self._environments = environments
        self._created_date = created_date
        self._updated_date = updated_date

    @classmethod
    def from_dict(cls, dikt) -> 'VersionedModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The VersionedModel of this VersionedModel.  # noqa: E501
        :rtype: VersionedModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this VersionedModel.


        :return: The name of this VersionedModel.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this VersionedModel.


        :param name: The name of this VersionedModel.
        :type name: str
        """

        self._name = name

    @property
    def id(self) -> str:
        """Gets the id of this VersionedModel.


        :return: The id of this VersionedModel.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this VersionedModel.


        :param id: The id of this VersionedModel.
        :type id: str
        """

        self._id = id

    @property
    def version(self) -> str:
        """Gets the version of this VersionedModel.


        :return: The version of this VersionedModel.
        :rtype: str
        """
        return self._version

    @version.setter
    def version(self, version: str):
        """Sets the version of this VersionedModel.


        :param version: The version of this VersionedModel.
        :type version: str
        """

        self._version = version

    @property
    def tags(self) -> List[str]:
        """Gets the tags of this VersionedModel.


        :return: The tags of this VersionedModel.
        :rtype: List[str]
        """
        return self._tags

    @tags.setter
    def tags(self, tags: List[str]):
        """Sets the tags of this VersionedModel.


        :param tags: The tags of this VersionedModel.
        :type tags: List[str]
        """

        self._tags = tags

    @property
    def labels(self) -> List[str]:
        """Gets the labels of this VersionedModel.


        :return: The labels of this VersionedModel.
        :rtype: List[str]
        """
        return self._labels

    @labels.setter
    def labels(self, labels: List[str]):
        """Sets the labels of this VersionedModel.


        :param labels: The labels of this VersionedModel.
        :type labels: List[str]
        """

        self._labels = labels

    @property
    def environments(self) -> List[Environment]:
        """Gets the environments of this VersionedModel.


        :return: The environments of this VersionedModel.
        :rtype: List[Environment]
        """
        return self._environments

    @environments.setter
    def environments(self, environments: List[Environment]):
        """Sets the environments of this VersionedModel.


        :param environments: The environments of this VersionedModel.
        :type environments: List[Environment]
        """

        self._environments = environments

    @property
    def created_date(self) -> datetime:
        """Gets the created_date of this VersionedModel.


        :return: The created_date of this VersionedModel.
        :rtype: datetime
        """
        return self._created_date

    @created_date.setter
    def created_date(self, created_date: datetime):
        """Sets the created_date of this VersionedModel.


        :param created_date: The created_date of this VersionedModel.
        :type created_date: datetime
        """

        self._created_date = created_date

    @property
    def updated_date(self) -> datetime:
        """Gets the updated_date of this VersionedModel.


        :return: The updated_date of this VersionedModel.
        :rtype: datetime
        """
        return self._updated_date

    @updated_date.setter
    def updated_date(self, updated_date: datetime):
        """Sets the updated_date of this VersionedModel.


        :param updated_date: The updated_date of this VersionedModel.
        :type updated_date: datetime
        """

        self._updated_date = updated_date